New: Actualizar Script Install/Update para que copia dinámicamente los directorios dentro de src.
New: Actualizar Script Uninstall para que no borre por defecto la carpeta de configuración /etc/watchful.
New: Mover la configuración de los módulos a modules.json.
Update: Mover la configuración de telegram a un subnodo llamado telegram.
New: Crear una función read_config en ModuleMase para leer la configuración del módulo.
Fix: Añadir comprobación de si existe el archivo de configuración que deseamos leer, y así evitar la excepción.

+++ uninstall.sh
@@ -1,12 +1,41 @@
 #!/bin/bash
+
+CLEAR_ALL=NO
+
+help_usage() {
+    echo "usage: uninstall [[-a | --all] | [-h | --help]]"
+}
+
+while [[ $# -gt 0 ]]; do
+  key="$1"
+  case $key in
+    -a | --all )
+      CLEAR_ALL=YES
+      ;;
+    -h | --help )
+      help_usage
+      exit
+      ;;
+    * )
+      echo "Error: Unknow parameter (${key})!!"
+      help_usage
+      exit 1
+    esac
+    shift
+done
+
 if [ "$(id -u)" != "0" ]; then
     echo "Sorry, you are not root."
     exit 1
 fi
+
 systemctl disable watchful.timer
 systemctl stop watchful.timer
-rm /lib/systemd/system/watchful.service
-rm /lib/systemd/system/watchful.timer
+rm '/lib/systemd/system/watchful.service'
+rm '/lib/systemd/system/watchful.timer'
 systemctl daemon-reload
-rm -rf '/etc/watchful'
 rm -rf '/opt/watchful'
+
+if [[ "${CLEAR_ALL}" == "YES" ]]; then
+  rm -rf '/etc/watchful'
+fi
